Move `yaml-util.h` to `yaml_util.h`.
