+++ plugin/src/Machinery.cc
@@ -23,7 +23,7 @@
 #include "txn_box/Config.h"
 #include "txn_box/Context.h"
 
-#include "txn_box/yaml-util.h"
+#include "txn_box/yaml_util.h"
 #include "txn_box/ts_util.h"
 
 using swoc::TextView;
