+++ plugin/src/txn_box.cc
@@ -29,7 +29,7 @@
 #include "txn_box/Context.h"
 
 #include "txn_box/ts_util.h"
-#include "txn_box/yaml-util.h"
+#include "txn_box/yaml_util.h"
 
 using swoc::TextView;
 using swoc::Errata;
